Fix release builds: real updater pubkey, Linux timeval, Windows Npcap.
Embed the minisign public key in tauri.conf for updater signing, align Linux pcap timeval types with libc on Linux CI, and make Chocolatey Npcap install non-fatal when the package is unavailable.

Made-with: Cursor

+++ src-tauri/src/packet_capture/pcap_reader.rs
@@ -239,7 +239,7 @@ impl MmapPcapReader {
         let pcap_header = pcap::PacketHeader {
             ts: libc::timeval {
                 tv_sec: header.ts_sec as i64,
-                tv_usec: ts_usec as i32,
+                tv_usec: ts_usec as i64,
             },
             caplen: header.caplen,
             len: header.len,
